# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CofradeNet API
  description: |
    API REST de la plataforma CofradeNet. Gestiona hermandades, bandas, procesiones,
    usuarios cofrades y el panel de administración de la plataforma.

    ## Autenticación
    La mayoría de los endpoints de escritura requieren un token JWT Bearer.
    Obtén el token con `POST /auth/login` e inclúyelo en la cabecera:
    `Authorization: Bearer <token>`

    ## Roles disponibles
    | Rol | Descripción |
    |-----|-------------|
    | `cofrade` | Usuario registrado. Puede seguir entidades, crear publicaciones y gestionar su perfil. |
    | `hermandad` | Administra el perfil de su hermandad y sus procesiones. |
    | `banda` | Administra el perfil de su banda y su agenda de actuaciones. |
    | `admin` | Acceso completo a la plataforma. |
  version: 1.0.0
  contact:
    name: Equipo CofradeNet
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /procesiones:
    get:
      tags: [procesiones]
      summary: Listar todas las procesiones
      operationId: listarProcesiones
      security: []
      responses:
        '200':
          description: Lista de procesiones ordenada por fecha.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Procesion'
    post:
      tags: [procesiones]
      summary: Crear procesión (HERMANDAD o ADMIN)
      operationId: crearProcesion
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProcesionRequest'
      responses:
        '201':
          description: Procesión creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Procesion'
        '403':
          $ref: '#/components/responses/Error403'
  /procesiones/buscar:
    get:
      tags: [procesiones]
      summary: Buscar procesiones con filtros
      description: |
        Permite filtrar procesiones por ciudad, día de la semana, nombre, hermandad o banda.
        Al menos uno de los parámetros debe estar presente.
      operationId: buscarProcesiones
      security: []
      parameters:
        - name: ciudad
          in: query
          schema:
            type: string
          example: "Sevilla"
        - name: dia
          in: query
          schema:
            type: string
          example: "Viernes Santo"
        - name: nombre
          in: query
          schema:
            type: string
        - name: hermandad
          in: query
          schema:
            type: string
        - name: banda
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Procesiones que coinciden con los filtros.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Procesion'
  /procesiones/hermandad/{id}:
    get:
      tags: [procesiones]
      summary: Procesiones de una hermandad
      operationId: procesionesPorHermandad
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Procesiones de la hermandad indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Procesion'
  /procesiones/{id}:
    get:
      tags: [procesiones]
      summary: Obtener procesión por ID
      operationId: getProcesion
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Procesión con itinerario (puntos de mapa) ordenado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Procesion'
        '404':
          $ref: '#/components/responses/Error404'
    delete:
      tags: [procesiones]
      summary: Eliminar procesión (HERMANDAD propietaria o ADMIN)
      operationId: eliminarProcesion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Procesión eliminada.
        '403':
          $ref: '#/components/responses/Error403'
  /procesiones/{id}/ficha/{anio}:
    get:
      tags: [procesiones]
      summary: Ficha completa de una procesión para un año
      description: |
        Devuelve la procesión enriquecida con el itinerario oficial y las bandas
        participantes del año indicado.
      operationId: fichaAnual
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: anio
          in: path
          required: true
          schema:
            type: integer
            example: 2025
      responses:
        '200':
          description: Ficha anual completa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FichaAnual'
        '404':
          $ref: '#/components/responses/Error404'
  /procesiones/{id}/participaciones:
    get:
      tags: [procesiones]
      summary: Listar bandas participantes en la procesión
      operationId: listarParticipaciones
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de participaciones ordenada por año descendente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Participacion'
    post:
      tags: [procesiones]
      summary: Añadir banda a la procesión (HERMANDAD o ADMIN)
      operationId: addParticipacion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateParticipacionRequest'
      responses:
        '201':
          description: Participación registrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participacion'
  /procesiones/{id}/participaciones/{pid}:
    put:
      tags: [procesiones]
      summary: Actualizar participación (HERMANDAD o ADMIN)
      operationId: updateParticipacion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateParticipacionRequest'
      responses:
        '200':
          description: Participación actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participacion'
    delete:
      tags: [procesiones]
      summary: Eliminar participación (HERMANDAD o ADMIN)
      operationId: deleteParticipacion
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Participación eliminada.
  /procesiones/{id}/itinerario:
    get:
      tags: [procesiones]
      summary: Obtener itinerarios anuales de la procesión
      operationId: getItinerarios
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lista de itinerarios anuales (horarios y recorrido).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Itinerario'
    post:
      tags: [procesiones]
      summary: Publicar itinerario de un año (HERMANDAD o ADMIN)
      operationId: crearItinerario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItinerarioRequest'
      responses:
        '201':
          description: Itinerario creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerario'
    put:
      tags: [procesiones]
      summary: Actualizar itinerario existente (HERMANDAD o ADMIN)
      description: El cuerpo debe incluir `itinerarioId` para identificar cuál actualizar.
      operationId: updateItinerario
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItinerarioRequest'
      responses:
        '200':
          description: Itinerario actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerario'
  /procesiones/{id}/pasos:
    post:
      tags: [procesiones]
      summary: Añadir paso a la procesión (HERMANDAD o ADMIN)
      operationId: crearPaso
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePasoRequest'
      responses:
        '201':
          description: Paso creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paso'
  /procesiones/{id}/pasos/{pasoId}:
    put:
      tags: [procesiones]
      summary: Actualizar paso (HERMANDAD o ADMIN)
      operationId: updatePaso
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pasoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasoRequest'
      responses:
        '200':
          description: Paso actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Paso'
    delete:
      tags: [procesiones]
      summary: Eliminar paso (HERMANDAD o ADMIN)
      operationId: deletePaso
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: pasoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Paso eliminado.

  # ── SEGUIMIENTOS ──────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en `POST /auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador numérico del recurso.
  responses:
...
    Error403:
      description: Sin permisos — el usuario no tiene el rol necesario.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Hermandad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        nombrePopular:
          type: string
        titulares:
          type: array
          items:
            type: string
        añoFundacion:
          type: integer
          nullable: true
        templo:
          type: string
        diaSalida:
          type: string
          nullable: true
        descripcion:
          type: string
          nullable: true
        imagenEscudo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'

    # ── Bandas ────────────────────────────────
...
    Banda:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        estiloMusical:
          type: string
        historia:
          type: string
          nullable: true
        numeroComponentes:
          type: integer
          nullable: true
        imagenLogo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'
        repertorio:
          type: array
          items:
            $ref: '#/components/schemas/Marcha'
        eventos:
          type: array
          items:
            $ref: '#/components/schemas/Evento'

    # ── Procesiones ───────────────────────────
    CreateProcesionRequest:
      type: object
      required: [nombre, diaSemana, fecha, horaSalida, hermandadId]
      properties:
        nombre:
          type: string
          example: "Estación de Penitencia 2025"
        diaSemana:
          type: string
          example: "Madrugada"
        fecha:
          type: string
          format: date
          example: "2025-04-18"
        horaSalida:
          type: string
          example: "00:30:00"
        horaEntrada:
          type: string
          nullable: true
        hermandadId:
          type: integer
    Procesion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        diaSemana:
          type: string
        fecha:
          type: string
          format: date
        horaSalida:
          type: string
        horaEntrada:
          type: string
          nullable: true
        hermandad:
          $ref: '#/components/schemas/Hermandad'
        itinerario:
          type: array
          items:
            $ref: '#/components/schemas/PuntoItinerario'
          description: Puntos GPS del recorrido ordenados por `orden`.
    PuntoItinerario:
      type: object
      properties:
        id:
          type: integer
        orden:
          type: integer
        latitud:
          type: number
          format: double
        longitud:
          type: number
          format: double
    FichaAnual:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        hermandad:
          $ref: '#/components/schemas/Hermandad'
        itinerarios:
          type: array
          items:
            $ref: '#/components/schemas/Itinerario'
        participaciones:
          type: array
          items:
            $ref: '#/components/schemas/Participacion'

    # ── Participaciones ───────────────────────
    CreateParticipacionRequest:
      type: object
      required: [bandaId, anio]
      properties:
        bandaId:
          type: integer
        anio:
          type: integer
          example: 2025
        ubicacion:
          type: string
          nullable: true
          example: "Paso de Cristo"
    UpdateParticipacionRequest:
      type: object
      properties:
        anio:
          type: integer
        ubicacion:
          type: string
    Participacion:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        anio:
          type: integer
        ubicacion:
          type: string
          nullable: true
        banda:
          $ref: '#/components/schemas/Banda'

    # ── Itinerarios ───────────────────────────
    CreateItinerarioRequest:
      type: object
      required: [anio]
      properties:
        anio:
          type: integer
          example: 2025
        horarioSalida:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "00:30:00"
        horarioEntrada:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
          example: "08:45:00"
        recorrido:
          type: string
          example: "Iglesia → C/ Sierpes → Plaza de San Francisco → Catedral"
    UpdateItinerarioRequest:
      type: object
      required: [itinerarioId]
      properties:
        itinerarioId:
          type: integer
          description: ID del itinerario a actualizar.
        horarioSalida:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
        horarioEntrada:
          type: string
          pattern: '^\d{2}:\d{2}:\d{2}$'
        recorrido:
          type: string
    Itinerario:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        anio:
          type: integer
        horarioSalida:
          type: string
          nullable: true
        horarioEntrada:
          type: string
          nullable: true
        recorrido:
          type: string
          nullable: true

    # ── Pasos ─────────────────────────────────
    CreatePasoRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          example: "Santísimo Cristo de la Buena Muerte"
        tipo:
          type: string
          example: "Cristo"
        orden:
          type: integer
          example: 1
        descripcion:
          type: string
    UpdatePasoRequest:
      type: object
      properties:
        nombre:
          type: string
        tipo:
          type: string
        orden:
          type: integer
        descripcion:
          type: string
    Paso:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        tipo:
          type: string
          nullable: true
        orden:
          type: integer
          nullable: true
        descripcion:
          type: string
          nullable: true

    # ── Eventos ───────────────────────────────
...
    Evento:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        fechaHora:
          type: string
          format: date-time
        lugar:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
        bandaId:
          type: integer
...
    Ciudad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        provincia:
          type: string
        pais:
          type: string
...
    Marcha:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        compositor:
          type: string
          nullable: true

    # ── Admin ─────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: "La procesión no existe"
        error:
          type: string
          example: "Not Found"
security:
  - BearerAuth: []